more toJS fixes
